waste_classification_app/requirements.txt
streamlit == 1.50.0
tensorflow-cpu==2.20.0
numpy
pillow